+++ winrsdk/src/main/kotlin/com/avafli/winrsdk/domain/PrizeClaim.kt
@@ -116,7 +116,8 @@ internal data class PrizeClaimForm(
 
         val usStates = listOf(
             "Alabama", "Alaska", "Arizona", "Arkansas", "California", "Colorado",
-            "Connecticut", "Delaware", "Florida", "Georgia", "Hawaii", "Idaho",
+            "Connecticut", "Delaware", "District of Columbia", "Florida",
+            "Georgia", "Hawaii", "Idaho",
             "Illinois", "Indiana", "Iowa", "Kansas", "Kentucky", "Louisiana",
             "Maine", "Maryland", "Massachusetts", "Michigan", "Minnesota",
             "Mississippi", "Missouri", "Montana", "Nebraska", "Nevada",